Add meter mock collector and validate

+++ validator/src/main/java/org/apache/skywalking/plugin/test/agent/tool/validator/assertor/element/GreetEqualAssertor.java
@@ -27,7 +27,7 @@ public class GreetEqualAssertor extends ElementAssertor {
 
     @Override
     public void assertValue(String desc, String actualValue) {
-        if (Long.parseLong(actualValue) < Long.parseLong(exceptedValue)) {
+        if (Double.parseDouble(actualValue) < Double.parseDouble(exceptedValue)) {
             throw new ValueAssertFailedException(desc, " ge " + exceptedValue, actualValue);
         }
     }
